Added circle of radius 200 with centre 300,300. Changed code in function for integer division.

+++ Ex2.4/functions.cpp
@@ -3,7 +3,7 @@
 
 void makePolygonCircle(SPA::Window&window,int sides,int radius,int x,int y){
 
-    int a = (180*(sides-2))/sides;
+    float a = (180*((float)sides-2))/sides;
     float b = (M_PI/180)*(a/2);
     float h = cos(b)*radius; //HALF SIDE LENGTH
 
@@ -16,7 +16,7 @@ void makePolygonCircle(SPA::Window&window,int sides,int radius,int x,int y){
 
     for (int i = 0; i <(sides -1); ++i) {
 
-        window.rotateDegrees(360/sides);
+        window.rotateDegrees(360./sides);
         window.forward(h);
     }
     window.stopLine();
